+++ src/posts/posts.controller.ts
@@ -1,34 +1,44 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete, UsePipes } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, UseGuards, UnauthorizedException } from '@nestjs/common';
 import { PostsService } from './posts.service';
 import { CreatePostDto } from './dto/create-post.dto';
 import { UpdatePostDto } from './dto/update-post.dto';
+import { JWTGuard } from 'src/auth/guard/jwt.guard';
+import { UserDecorator } from 'src/common/decorators/user.decorator';
+import { JwtPayloadDto } from 'src/auth/dto/jwt-payload.dto';
+import { RoleEnum } from 'src/common/enum/roles.enum';
 
 @Controller('posts')
 export class PostsController {
-  constructor(private readonly postsService: PostsService) {}
+    constructor(private readonly postsService: PostsService) {}
 
-  @Post()
-  create(@Body() createPostDto: CreatePostDto) {
-    return this.postsService.create(createPostDto);
-  }
+    @Post()
+    @UseGuards(JWTGuard)
+    create(@UserDecorator() user: JwtPayloadDto, @Body() createPostDto: CreatePostDto) {
+        if (user.role == RoleEnum.STUDENT) throw new UnauthorizedException();
+        return this.postsService.create(createPostDto);
+    }
 
-  @Get()
-  findAll() {
-    return this.postsService.findAll();
-  }
+    @Get()
+    findAll() {
+        return this.postsService.findAll();
+    }
 
-  @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.postsService.findOne(id);
-  }
+    @Get(':id')
+    findOne(@Param('id') id: string) {
+        return this.postsService.findOne(id);
+    }
 
-  @Patch(':id')
-  update(@Param('id') id: string, @Body() updatePostDto: UpdatePostDto) {
-    return this.postsService.update(id, updatePostDto);
-  }
+    @Patch(':id')
+    @UseGuards(JWTGuard)
+    update(@UserDecorator() user: JwtPayloadDto, @Param('id') id: string, @Body() updatePostDto: UpdatePostDto) {
+        if (user.role == RoleEnum.STUDENT) throw new UnauthorizedException();
+        return this.postsService.update(id, updatePostDto);
+    }
 
-  @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.postsService.remove(id);
-  }
+    @Delete(':id')
+    @UseGuards(JWTGuard)
+    remove(@UserDecorator() user: JwtPayloadDto, @Param('id') id: string) {
+        if (user.role == RoleEnum.STUDENT) throw new UnauthorizedException();
+        return this.postsService.remove(id);
+    }
 }
