+++ src/subject/subject.controller.ts
@@ -1,34 +1,46 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, UseGuards, UnauthorizedException } from '@nestjs/common';
 import { SubjectService } from './subject.service';
 import { CreateSubjectDto } from './dto/create-subject.dto';
 import { UpdateSubjectDto } from './dto/update-subject.dto';
+import { JWTGuard } from 'src/auth/guard/jwt.guard';
+import { UserDecorator } from 'src/common/decorators/user.decorator';
+import { JwtPayloadDto } from 'src/auth/dto/jwt-payload.dto';
+import { RoleEnum } from 'src/common/enum/roles.enum';
 
 @Controller('subject')
 export class SubjectController {
-  constructor(private readonly subjectService: SubjectService) {}
+    constructor(private readonly subjectService: SubjectService) {}
 
-  @Post()
-  create(@Body() createSubjectDto: CreateSubjectDto) {
-    return this.subjectService.create(createSubjectDto);
-  }
+    @Post()
+    @UseGuards(JWTGuard)
+    create(@UserDecorator() user: JwtPayloadDto, @Body() createSubjectDto: CreateSubjectDto) {
+        if (user.role == RoleEnum.STUDENT) throw new UnauthorizedException();
+        return this.subjectService.create(createSubjectDto);
+    }
 
-  @Get()
-  findAll() {
-    return this.subjectService.findAll();
-  }
+    @Get()
+    findAll() {
+        return this.subjectService.findAll();
+    }
 
-  @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.subjectService.findOne(id);
-  }
+    @Get(':id')
+    findOne(@Param('id') id: string) {
+        return this.subjectService.findOne(id);
+    }
 
-  @Patch(':id')
-  update(@Param('id') id: string, @Body() updateSubjectDto: UpdateSubjectDto) {
-    return this.subjectService.update(id, updateSubjectDto);
-  }
+    @Patch(':id')
+    @UseGuards(JWTGuard)
+    update(@UserDecorator() user: JwtPayloadDto, @Param('id') id: string, @Body() updateSubjectDto: UpdateSubjectDto) {
+        if (user.role == RoleEnum.STUDENT) throw new UnauthorizedException();
 
-  @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.subjectService.remove(id);
-  }
+        return this.subjectService.update(id, updateSubjectDto);
+    }
+
+    @Delete(':id')
+    @UseGuards(JWTGuard)
+    remove(@UserDecorator() user: JwtPayloadDto, @Param('id') id: string) {
+        if (user.role == RoleEnum.STUDENT) throw new UnauthorizedException();
+
+        return this.subjectService.remove(id);
+    }
 }
